update refactoring list component
